worker_service/requirements.txt
# Celery
celery==5.4.0

# OpenCV + NumPy
opencv-python==4.7.0.72
numpy==1.24.2

# SQLAlchemy, если воркер лезет в ту же БД
SQLAlchemy==2.0.5.post1
psycopg2-binary==2.9.5

# Чтение переменных окружения
python-dotenv==1.0.0

# (Опционально) httpx, если воркер берёт video_path из API через HTTP
httpx==0.23.3

requests==2.28.2

flake8==6.0.0

mypy==1.1.1

graypy==2.1.0
